Support Data Diagnose and personalized recommendation switch.

+++ aliyun-java-sdk-airec/src/main/java/com/aliyuncs/airec/model/v20201126/DescribeSyncReportDetailRequest.java
@@ -29,9 +29,9 @@ public class DescribeSyncReportDetailRequest extends RoaAcsRequest<DescribeSyncR
 
 	private String levelType;
 
-	private Long endTime;
+	private Integer endTime;
 
-	private Long startTime;
+	private Integer startTime;
 
 	private String type;
 	public DescribeSyncReportDetailRequest() {
@@ -66,22 +66,22 @@ public class DescribeSyncReportDetailRequest extends RoaAcsRequest<DescribeSyncR
 		}
 	}
 
-	public Long getEndTime() {
+	public Integer getEndTime() {
 		return this.endTime;
 	}
 
-	public void setEndTime(Long endTime) {
+	public void setEndTime(Integer endTime) {
 		this.endTime = endTime;
 		if(endTime != null){
 			putQueryParameter("endTime", endTime.toString());
 		}
 	}
 
-	public Long getStartTime() {
+	public Integer getStartTime() {
 		return this.startTime;
 	}
 
-	public void setStartTime(Long startTime) {
+	public void setStartTime(Integer startTime) {
 		this.startTime = startTime;
 		if(startTime != null){
 			putQueryParameter("startTime", startTime.toString());
