+++ aliyun-java-sdk-airec/src/main/java/com/aliyuncs/airec/model/v20201126/ListDashboardMetricsFlowsRequest.java
@@ -29,9 +29,9 @@ public class ListDashboardMetricsFlowsRequest extends RoaAcsRequest<ListDashboar
 
 	private String instanceId;
 
-	private Long endTime;
+	private Integer endTime;
 
-	private Long startTime;
+	private Integer startTime;
 	public ListDashboardMetricsFlowsRequest() {
 		super("Airec", "2020-11-26", "ListDashboardMetricsFlows", "airec");
 		setUriPattern("/v2/openapi/instances/[instanceId]/dashboard/metrics/flows");
@@ -64,22 +64,22 @@ public class ListDashboardMetricsFlowsRequest extends RoaAcsRequest<ListDashboar
 		}
 	}
 
-	public Long getEndTime() {
+	public Integer getEndTime() {
 		return this.endTime;
 	}
 
-	public void setEndTime(Long endTime) {
+	public void setEndTime(Integer endTime) {
 		this.endTime = endTime;
 		if(endTime != null){
 			putQueryParameter("endTime", endTime.toString());
 		}
 	}
 
-	public Long getStartTime() {
+	public Integer getStartTime() {
 		return this.startTime;
 	}
 
-	public void setStartTime(Long startTime) {
+	public void setStartTime(Integer startTime) {
 		this.startTime = startTime;
 		if(startTime != null){
 			putQueryParameter("startTime", startTime.toString());
