+++ aliyun-java-sdk-airec/src/main/java/com/aliyuncs/airec/model/v20201126/ListRulesRequest.java
@@ -33,11 +33,11 @@ public class ListRulesRequest extends RoaAcsRequest<ListRulesResponse> {
 
 	private String sceneId;
 
-	private Long endTime;
+	private Integer endTime;
 
 	private Integer page;
 
-	private Long startTime;
+	private Integer startTime;
 
 	private String status;
 	public ListRulesRequest() {
@@ -94,11 +94,11 @@ public class ListRulesRequest extends RoaAcsRequest<ListRulesResponse> {
 		}
 	}
 
-	public Long getEndTime() {
+	public Integer getEndTime() {
 		return this.endTime;
 	}
 
-	public void setEndTime(Long endTime) {
+	public void setEndTime(Integer endTime) {
 		this.endTime = endTime;
 		if(endTime != null){
 			putQueryParameter("endTime", endTime.toString());
@@ -116,11 +116,11 @@ public class ListRulesRequest extends RoaAcsRequest<ListRulesResponse> {
 		}
 	}
 
-	public Long getStartTime() {
+	public Integer getStartTime() {
 		return this.startTime;
 	}
 
-	public void setStartTime(Long startTime) {
+	public void setStartTime(Integer startTime) {
 		this.startTime = startTime;
 		if(startTime != null){
 			putQueryParameter("startTime", startTime.toString());
